SpringDemo/src/main/java/br/com/uri/SpringDemo/controller/HelloController.java
package br.com.uri.SpringDemo.controller;

import br.com.uri.SpringDemo.service.HelloService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloController {

    private final HelloService helloService;

    public HelloController(HelloService helloService) {
        this.helloService = helloService;
    }

    @GetMapping("/hello")
    public ResponseEntity<String> sayHello() {
        return ResponseEntity.ok(helloService.sayHello());
    }
}
